requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.31.0
pydantic[email]==2.6.0
pydantic-settings==2.1.0
passlib==1.7.4
python-jose[cryptography]==3.3.0
redis==5.0.1
